requirement.txt
flask
flask-cors
nltk
pdfminer.six
summa
fuzzywuzzy
python-Levenshtein
